Terminado el CRUD
ya se puede editar todos los documentos de  las auditorias

+++ application/views/admin/documentos/resumen/lista.php
@@ -54,15 +54,26 @@
                                     </center>
                                 </td>
                                 ';
-                            
-                            echo '	<td>
-                                <center>
-                                    <a href="'.base_url().'index.php/admin/documentos/resumen_eliminar/'.$row->IdResumen.'">
-                                        <img src="'.base_url().'includes/img/icons/eliminar.png" />
-                                    </a>
-                                </center>
-                            </td>
-                            ';
+                            if($row->Estado == 1){
+                                echo '	<td>
+                                    <center>
+                                        <a href="'.base_url().'index.php/admin/documentos/resumen_eliminar/'.$row->IdResumen.'">
+                                            <img src="'.base_url().'includes/img/icons/eliminar.png" />
+                                        </a>
+                                    </center>
+                                </td>
+                                ';
+                            }else{
+                                echo '	<td>
+                                    <center>
+                                        <a href="'.base_url().'index.php/admin/documentos/resumen_cambiar/'.$row->IdResumen.'">
+                                            <img src="'.base_url().'includes/img/icons/activar.png" />
+                                        </a>
+                                    </center>
+                                </td>
+                                ';
+                            }
+            
                             echo '</tr>';
                         }
                     }else{
